backend/requirements.txt
fastapi
uvicorn
requests
dnspython
shodan
python-whois
pydantic
python-nmap
motor
playwright
certifi
